+++ examples/resources/cloudtemple_compute_virtual_machine/resource.tf
@@ -10,8 +10,9 @@ data "cloudtemple_compute_datastore_cluster" "koukou" {
   name = "sdrs001-LIVE_KOUKOU"
 }
 
-resource "cloudtemple_compute_virtual_machine" "foo" {
-  name = "test-terraform-example-vm"
+# Deploying a new virtual machine with a given operating system
+resource "cloudtemple_compute_virtual_machine" "scratch" {
+  name = "from-scratch"
 
   memory                 = 8 * 1024 * 1024 * 1024
   cpu                    = 2
@@ -25,6 +26,51 @@ resource "cloudtemple_compute_virtual_machine" "foo" {
   datastore_cluster_id         = data.cloudtemple_compute_datastore_cluster.koukou.id
   guest_operating_system_moref = "amazonlinux2_64Guest"
 
+
+  tags = {
+    created_by = "Terraform"
+  }
+}
+
+# Clone an already existing virtual machine
+resource "cloudtemple_compute_virtual_machine" "cloned" {
+  name = "cloned"
+
+  clone_virtual_machine_id = cloudtemple_compute_virtual_machine.scratch.id
+
+  virtual_datacenter_id = data.cloudtemple_compute_virtual_datacenter.dc.id
+  host_cluster_id       = data.cloudtemple_compute_host_cluster.flo.id
+  datastore_cluster_id  = data.cloudtemple_compute_datastore_cluster.koukou.id
+
+  tags = {
+    created_by = "Terraform"
+  }
+}
+
+# Deploy an item from the content library
+data "cloudtemple_compute_content_library" "public" {
+  name = "PUBLIC"
+}
+
+data "cloudtemple_compute_content_library_item" "centos" {
+  content_library_id = data.cloudtemple_compute_content_library.public.id
+  name               = "20211115132417_master_linux-centos-8"
+}
+
+data "cloudtemple_compute_datastore" "ds" {
+  name = "ds001-bob-svc1-data4-eqx6"
+}
+
+resource "cloudtemple_compute_virtual_machine" "content-library" {
+  name = "from-content-library-item"
+
+  content_library_id      = data.cloudtemple_compute_content_library.public.id
+  content_library_item_id = data.cloudtemple_compute_content_library_item.centos.id
+
+  virtual_datacenter_id = data.cloudtemple_compute_virtual_datacenter.dc.id
+  host_cluster_id       = data.cloudtemple_compute_host_cluster.flo.id
+  datastore_cluster_id  = data.cloudtemple_compute_datastore_cluster.koukou.id
+  datastore_id          = data.cloudtemple_compute_datastore.ds.id
 
   tags = {
     created_by = "Terraform"