+++ CHANGELOG.md
@@ -4,10 +4,11 @@ NEW FEATURES:
 
   * The `cloudtemple_compute_content_library_item` datasource can now be used to read an item from the given content library.
   * The `cloudtemple_compute_content_library_items` datasource can now be used to read all items in a given content library.
+  * The `cloudtemple_compute_virtual_machine` resource now supports deploying a new virtual machine from a content library item.
 
 IMPROVEMENTS:
 
-  * The provider now periodically logs information regarding the state of the activity or job running while waiting for them to complete
+  * The provider now periodically logs information regarding the state of the activity or job running while waiting for them to complete.
 
 
 ## 0.2.2 (November 24, 2022)